requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anthropic==0.29.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
executing==2.0.1
filelock==3.15.3
fonttools==4.51.0
fsspec==2024.6.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jiter==0.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keyboard==0.13.5
kiwisolver==1.4.5
macholib==1.16.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.30.1
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pygame==2.5.2
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pynput==1.7.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyttsx3==2.90
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.31.0
rsa==4.9
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
stack-data==0.6.3
tiktoken==0.7.0
tk==0.1.0
tokenizers==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
PyQt5
PyQt5-Qt5
PyQt5-sip
screeninfo
python-dotenv
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
